Finished Assignment
Added the detail activity for each of the items, updated myAdapter, and change the style of the App. There are 10 items now instead of 5. When you enter the detail activity by clicking an item of the recycler view in main activity, the description, picture, title, and a toast message of the genre appears. The app allows you to look at a detail activity and go back to the main activity (by clicking the Up functionality). This app implements RecyclerView, RecyclerView.Adapter, RecyclerView.ViewHolder, and View.OnClickListener.